+++ src/main/java/gtclassic/common/tile/GTTileMagicEnergyConverter.java
@@ -2,8 +2,10 @@ package gtclassic.common.tile;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 import gtclassic.api.helpers.GTHelperFluid;
+import gtclassic.api.interfaces.IGTDebuggableTile;
 import gtclassic.api.material.GTMaterial;
 import gtclassic.api.material.GTMaterialGen;
 import gtclassic.api.recipe.GTRecipeMultiInputList;
@@ -56,7 +58,7 @@ import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
 public class GTTileMagicEnergyConverter extends TileEntityMachine
-		implements ITankListener, ITickable, IHasGui, IEUStorage, IEmitterTile, IEnergySourceInfo {
+		implements ITankListener, ITickable, IHasGui, IEUStorage, IEmitterTile, IEnergySourceInfo, IGTDebuggableTile {
 
 	@NetworkField(index = 4)
 	public int storage = 0;
@@ -66,6 +68,7 @@ public class GTTileMagicEnergyConverter extends TileEntityMachine
 	int slotInput = 0;
 	int slotOutput = 1;
 	int slotDisplay = 2;
+	int fuel = 0;
 	boolean enet = false;
 	public static final GTRecipeMultiInputList RECIPE_LIST = new GTRecipeMultiInputList("gt.magicfuels");
 
@@ -73,7 +76,7 @@ public class GTTileMagicEnergyConverter extends TileEntityMachine
 		super(3);
 		this.tank = new IC2Tank(1000);
 		this.tank.addListener(this);
-		this.addGuiFields("tank");
+		this.addGuiFields("tank", "fuel");
 	}
 
 	@Override
@@ -96,6 +99,10 @@ public class GTTileMagicEnergyConverter extends TileEntityMachine
 	public void onTankChanged(IFluidTank tank) {
 		this.getNetwork().updateTileGuiField(this, "tank");
 		this.inventory.set(2, ItemDisplayIcon.createWithFluidStack(this.tank.getFluid()));
+		if (this.tank.getFluidAmount() == 0) {
+			this.tank.setFluid(null);
+			this.setStackInSlot(slotDisplay, ItemStack.EMPTY);
+		}
 	}
 
 	@Override
@@ -103,6 +110,7 @@ public class GTTileMagicEnergyConverter extends TileEntityMachine
 		super.readFromNBT(nbt);
 		this.tank.readFromNBT(nbt.getCompoundTag("tank"));
 		this.storage = nbt.getInteger("storage");
+		this.fuel = nbt.getInteger("fuel");
 	}
 
 	@Override
@@ -110,6 +118,7 @@ public class GTTileMagicEnergyConverter extends TileEntityMachine
 		super.writeToNBT(nbt);
 		this.tank.writeToNBT(this.getTag(nbt, "tank"));
 		nbt.setInteger("storage", this.storage);
+		nbt.setInteger("fuel", this.fuel);
 		return nbt;
 	}
 
@@ -168,75 +177,61 @@ public class GTTileMagicEnergyConverter extends TileEntityMachine
 
 	@Override
 	public void update() {
-		boolean hasPower = this.storage > 0;
-		if ((hasPower) != this.getActive()) {
-			this.setActive(hasPower);
-		}
 		GTHelperFluid.doFluidContainerThings(this, this.tank, slotInput, slotOutput);
-		this.emptyCheck();
-		processLiquidMagicFuel();
-		processSolidMagicFuel();
+		processFuelValue();
+		this.setActive(this.fuel > 0 && this.storage < this.maxStorage);
+		if (this.fuel == 0) {
+			processMagicFuel();
+		}
 	}
 
-	/** Logic for processing of magic liquids **/
-	public void processLiquidMagicFuel() {
-		if (this.tank.getFluid() != null && isMagicLiquidFuel(this.tank.getFluid().getFluid())) {
-			// This tries burn fuel, 3ml at a time to generate 24eu
-			if (this.storage + this.production <= this.maxStorage && this.tank.getFluidAmount() >= 3) {
-				this.tank.setFluid(new FluidStack(this.getContained(), this.tank.getFluid().amount - 3));
-				this.storage = this.storage + this.production;
+	private void processFuelValue() {
+		if (this.fuel - 1 >= 0 && this.storage + this.production <= this.maxStorage) {
+			this.fuel = this.fuel - 1;
+			this.storage = this.storage + this.production;
+			this.getNetwork().updateTileGuiField(this, "fuel");
+		}
+	}
+
+	/** Logic for processing of magic fuels **/
+	private void processMagicFuel() {
+		if (this.tank.getFluid() != null && this.tank.getFluidAmount() >= 1000) {
+			int newFuelFromFluid = getMagicFuelValue(this.tank.getFluid().getFluid());
+			if (newFuelFromFluid > 0 && newFuelFromFluid + this.fuel < Integer.MAX_VALUE) {
+				this.fuel = this.fuel + newFuelFromFluid;
+				this.tank.setFluid(new FluidStack(this.getContained(), this.tank.getFluid().amount - 1000));
 				this.onTankChanged(this.tank);
 				return;
 			}
-			// If there isnt enough, burn the remainder at 1ml to generate 8eu
-			if (this.storage + (this.production / 3) <= this.maxStorage && this.tank.getFluidAmount() > 0) {
-				this.tank.setFluid(new FluidStack(this.getContained(), this.tank.getFluid().amount - 1));
-				this.storage = this.storage + (this.production / 3);
-				this.onTankChanged(this.tank);
-			}
 		}
-	}
-
-	public void processSolidMagicFuel() {
 		if (!this.getStackInSlot(slotInput).isEmpty()) {
-			int energy = getMagicSolidFuelValue(this.getStackInSlot(slotInput));
-			if (energy > 0) {
-				int generate = (int) (energy * world.rand.nextFloat());
-				if (generate + this.storage <= this.maxStorage) {
-					this.storage = this.storage + generate;
-					this.getStackInSlot(slotInput).shrink(1);
-				}
+			int newFuelFromStack = getMagicFuelValue(this.getStackInSlot(slotInput));
+			if (newFuelFromStack > 0 && newFuelFromStack + this.fuel < Integer.MAX_VALUE) {
+				this.fuel = this.fuel + newFuelFromStack;
+				this.getStackInSlot(slotInput).shrink(1);
 			}
-		}
-	}
-
-	/** Helper method because this tile is weird af **/
-	public void emptyCheck() {
-		if (this.tank.getFluidAmount() == 0) {
-			this.tank.setFluid(null);
-			this.setStackInSlot(slotDisplay, ItemStack.EMPTY);
 		}
 	}
 
 	/** Checks for compatible fluids **/
-	public boolean isMagicLiquidFuel(Fluid fluid) {
+	public int getMagicFuelValue(Fluid fluid) {
 		if (RECIPE_LIST.getRecipeList().isEmpty()) {
-			return false;
+			return 0;
 		}
 		for (MultiRecipe map : RECIPE_LIST.getRecipeList()) {
 			IRecipeInput input = map.getInput(0);
 			if (input instanceof RecipeInputFluid) {
 				Fluid fluidinput = ((RecipeInputFluid) input).fluid.getFluid();
 				if (fluidinput == fluid) {
-					return true;
+					return map.getOutputs().getMetadata().getInteger("RecipeTime");
 				}
 			}
 		}
-		return false;
+		return 0;
 	}
 
 	/** Checks for compatible ItemStacks **/
-	public int getMagicSolidFuelValue(ItemStack stack) {
+	public int getMagicFuelValue(ItemStack stack) {
 		if (RECIPE_LIST.getRecipeList().isEmpty()) {
 			return 0;
 		}
@@ -312,6 +307,10 @@ public class GTTileMagicEnergyConverter extends TileEntityMachine
 		return this.production + 1;
 	}
 
+	public int getFuel() {
+		return this.fuel;
+	}
+
 	@Override
 	public boolean emitsEnergyTo(IEnergyAcceptor var1, EnumFacing facing) {
 		return true;
@@ -346,8 +345,8 @@ public class GTTileMagicEnergyConverter extends TileEntityMachine
 		addModRecipe("liquidantimatter"); // AbsyssalCraft
 		addModRecipe("mana_fluid"); // Wizardry
 		addModRecipe("nacre_fluid"); // Wizardry
-		addRecipe(GTMaterialGen.get(Items.NETHER_WART), 60);
-		addRecipe(GTMaterialGen.getIc2(Ic2Items.terraWart), 80);
+		addRecipe(GTMaterialGen.get(Items.NETHER_WART), 200);
+		addRecipe(GTMaterialGen.getIc2(Ic2Items.terraWart), 300);
 		addRecipe(GTMaterialGen.get(Items.ENDER_PEARL), 8000);
 		addRecipe("dustEnderPearl", 8000);
 		addRecipe(GTMaterialGen.get(Items.ENDER_EYE), 10000);
@@ -360,7 +359,8 @@ public class GTTileMagicEnergyConverter extends TileEntityMachine
 	}
 
 	public static IRecipeModifier[] value(int amount) {
-		return new IRecipeModifier[] { ModifierType.RECIPE_LENGTH.create(amount) };
+		int varAmount = amount > 24 ? amount : 24;
+		return new IRecipeModifier[] { ModifierType.RECIPE_LENGTH.create(varAmount / 24) };
 	}
 
 	public static void addModRecipe(String name) {
@@ -406,5 +406,10 @@ public class GTTileMagicEnergyConverter extends TileEntityMachine
 
 	public static void removeRecipe(String id) {
 		RECIPE_LIST.removeRecipe(id);
+	}
+
+	@Override
+	public void getData(Map<String, Boolean> data) {
+		data.put("Stored Fuel Value: " + this.fuel + "/" + this.fuel * 24 + " EU", true);
 	}
 }